Wire loss is factored into damage calculations now Changed damage energy to floats, it prevents issues with low energies (And makes IW easier :) )

+++ src/main/java/blusunrize/immersiveengineering/api/energy/wires/IImmersiveConnectable.java
@@ -112,7 +112,7 @@ public interface IImmersiveConnectable
 	 * @param amount The amount available from this source
 	 * @param consume Call this to consume the amount of energy in the parameter
 	 */
-	default void addAvailableEnergy(int amount, Consumer<Integer> consume)
+	default void addAvailableEnergy(float amount, Consumer<Float> consume)
 	{}
 	
 	/**
